*) fix for stupid mistake in new ppm-calc which caused decimal digits beeing written to seedinfo
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@2463 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/yacy/yacyPeerActions.java
@@ -128,7 +128,7 @@ public class yacyPeerActions {
             seedDB.mySeed.put(yacySeed.ISPEED, Long.toString(0));
         }else{
             //set the PPM
-	        seedDB.mySeed.put(yacySeed.ISPEED, Float.toString(Math.round((float)indexedcdiff / ((float)uptimediff/60f))));
+	        seedDB.mySeed.put(yacySeed.ISPEED, Long.toString(Math.round((float)indexedcdiff / ((float)uptimediff/60f))));
 		}
         if(uptime > 0)
         	sb.setConfig("totalPPM", Long.toString(indexedc / uptime));
